trying out the github tests

+++ services/summaries/tests/test_summaries.py
@@ -145,13 +145,18 @@ def test_remove_summary(test_app_with_db, monkeypatch):
     assert response.json() == {"id": summary_id, "user_id": user_id, "query": "Who was Madonna?"}
 
 
-# got this far
 def test_remove_summary_incorrect_id(test_app_with_db):
-    response = test_app_with_db.delete("/summaries/999/")
+    response = test_app_with_db.post(
+        "/users/", data=json.dumps({"username": "Jane Doe"})
+    )
+    sample_user = response.json()
+    user_id = sample_user["id"]
+
+    response = test_app_with_db.delete(f"/users/{user_id}/summaries/999/")
     assert response.status_code == 404
     assert response.json()["detail"] == "Summary not found"
 
-    response = test_app_with_db.delete("/summaries/0/")
+    response = test_app_with_db.delete(f"/users/{user_id}/summaries/0/")
     assert response.status_code == 422
     assert response.json() == {
         "detail": [
@@ -182,7 +187,6 @@ def test_update_summary(test_app_with_db, monkeypatch):
         f"/users/{user_id}/summaries/", data=json.dumps({"query": "Who was Charles Darwin?"})
     )
     summary_id = response.json()["id"]
-
     response = test_app_with_db.put(
         f"/users/{user_id}/summaries/{summary_id}/",
         data=json.dumps({"query": "Who was Charles Darwin?", "summary": "updated!"}),
